Modified Wallet Model

+++ src/DigitalWallet/Features/Wallet/Common/Wallet.cs
@@ -3,8 +3,8 @@ namespace DigitalWallet.Features.Wallet.Common;
 public class UserWallet
 {
     public Guid Id { get; set; }
-    public int Balance  { get; set; }
-    public int BlockBalance { get; set; }
+    public decimal Balance  { get; set; }
+    public decimal BlockBalance { get; set; }
     public DateTime CreatedOn { get; set; } = DateTime.Now;
